Add built-in session support to `arc.http.res` Remove `arc.http.res_fmt` shortcut, just use `.res` Use full `arc[pragma][method]()` invocations in tests
