+++ tests/conftest.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 import os
-import sys
 
 import boto3
 import pytest
