--- tests/test_http_res_fmt.py
@@ -1,1 +1,0 @@
-# TODO: implement tests when Arc's req/res fixtures can be ported to py; see: https://github.com/architect/req-res-fixtures
