+++ tests/test_queues.py
@@ -1,7 +1,6 @@
 # # -*- coding: utf-8 -*-
 import json
 from uuid import UUID
-
 import arc
